Move state functionality to separate part of API
