+++ src/EnergyPlus/api/func.h
@@ -51,6 +51,7 @@
 
 #include <EnergyPlus/api/TypeDefs.h>
 #include <EnergyPlus/api/EnergyPlusAPI.h>
+#include <EnergyPlus/api/state.h>
 
 #ifdef __cplusplus
 extern "C" {
