+++ src/EnergyPlus/api/runtime.cc
@@ -48,30 +48,10 @@
 #include <EnergyPlus/api/EnergyPlusPgm.hh>
 #include <EnergyPlus/DataGlobals.hh>
 #include <EnergyPlus/Data/EnergyPlusData.hh>
-#include <EnergyPlus/InputProcessing/IdfParser.hh>
-#include <EnergyPlus/InputProcessing/InputProcessor.hh>
-#include <EnergyPlus/InputProcessing/InputValidation.hh>
 #include <EnergyPlus/PluginManager.hh>
 #include <EnergyPlus/api/runtime.h>
-#include <EnergyPlus/StateManagement.hh>
 #include <EnergyPlus/UtilityRoutines.hh>
 #include <utility>
-
-EnergyPlusState stateNew() {
-    auto *state = new EnergyPlus::EnergyPlusData;
-    return reinterpret_cast<EnergyPlusState>(state);
-}
-
-void stateReset(EnergyPlusState state) {
-    auto *this_state = reinterpret_cast<EnergyPlus::EnergyPlusData *>(state);
-    EnergyPlus::clearAllStates(*this_state);
-    // also clear out the input processor since the clearAllStates does not do that.
-    EnergyPlus::inputProcessor = EnergyPlus::InputProcessor::factory();
-}
-
-void stateDelete(EnergyPlusState state) {
-    delete reinterpret_cast<EnergyPlus::EnergyPlusData *>(state);
-}
 
 int energyplus(EnergyPlusState state, int argc, const char *argv[]) {
 //    argv[0] = "energyplus";
