+++ src/EnergyPlus/api/state.cc
@@ -45,10 +45,26 @@
 // OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-#ifndef EnergyPlusBase_h_INCLUDED
-#define EnergyPlusBase_h_INCLUDED
+#include <EnergyPlus/api/state.h>
 
-typedef float Real32;  // Platform-specific: C++ has no defined precision floating point types
-typedef double Real64; // Platform-specific: C++ has no defined precision floating point types
+#include <EnergyPlus/Data/EnergyPlusData.hh>
+#include <EnergyPlus/InputProcessing/IdfParser.hh>
+#include <EnergyPlus/InputProcessing/InputProcessor.hh>
+#include <EnergyPlus/InputProcessing/InputValidation.hh>
+#include <EnergyPlus/StateManagement.hh>
 
-#endif
+EnergyPlusState stateNew() {
+    auto *state = new EnergyPlus::EnergyPlusData;
+    return reinterpret_cast<EnergyPlusState>(state);
+}
+
+void stateReset(EnergyPlusState state) {
+    auto *this_state = reinterpret_cast<EnergyPlus::EnergyPlusData *>(state);
+    EnergyPlus::clearAllStates(*this_state);
+    // also clear out the input processor since the clearAllStates does not do that.
+    EnergyPlus::inputProcessor = EnergyPlus::InputProcessor::factory();
+}
+
+void stateDelete(EnergyPlusState state) {
+    delete reinterpret_cast<EnergyPlus::EnergyPlusData *>(state);
+}
