+++ src/EnergyPlus/api/state.h
@@ -45,10 +45,54 @@
 // OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-#ifndef EnergyPlusBase_h_INCLUDED
-#define EnergyPlusBase_h_INCLUDED
+#ifndef EnergyPlusAPIState_h_INCLUDED
+#define EnergyPlusAPIState_h_INCLUDED
 
-typedef float Real32;  // Platform-specific: C++ has no defined precision floating point types
-typedef double Real64; // Platform-specific: C++ has no defined precision floating point types
+#include <EnergyPlus/api/EnergyPlusAPI.h>
+
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/// \brief This is typedef for an instance that stores the "state" of an EnergyPlus simulation.
+/// \details The state of an EnergyPlus simulation is held heavily in the global program state.  To alleviate issues
+///          with this design in the context of new applications and workflows, the global program state is being
+///          refactored into a single variable which will ultimately own and manage the entire state of the simulation.
+///          In order to keep track of the simulation during API calls, a reference to a state must be kept and
+///          managed by the API client.  The client must call `stateNew` to create a new state instance, and pass this
+///          to any functions that need a state instance.  Once a simulation is complete, the client can call
+///          `stateReset`, which will reset the state instance passed in, along with any global variables that are not
+///          yet refactored into this state instance.  Once complete, the client has the ability to free the memory of
+///          the state by using the `stateDelete` function..
+typedef void * EnergyPlusState;
+
+/// \brief Creates a new simulation state instance and returns it for the client to store while running simulations
+/// \details This function creates a new instance that is used in running simulations from the API.  The state created
+///          in this function is used by passing it into specific run functions.  When a simulation is complete, the
+///          state can be reset using the `stateReset` function, or deleted completely with the `stateDelete` function.
+/// \see EnergyPlusState
+/// \see stateReset
+/// \see stateDelete
+ENERGYPLUSLIB_API EnergyPlusState stateNew();
+/// \brief Resets the simulation state of EnergyPlus
+/// \details A simulation state is created by calling the `stateNew` function.  After a simulation is complete, if a
+///          second is to be run using the same memory space, the simulation state must be cleared with this function,
+///          or unexpected errors will occur.  Once the client is fully finished with the state, it can be deleted
+///          entirely with `stateDelete`.
+/// \param[in] state The simulation state of the simulation to clear.
+/// \remark This function will also clear any callback functions, so callback functions must be registered again.
+ENERGYPLUSLIB_API void stateReset(EnergyPlusState state);
+/// \brief Deletes a simulation state instance once the client is fully finished with it.
+/// \details A simulation state is created by calling the `stateNew` function.  After the client is finished, the memory
+///          can be reset by calling the `stateReset` function, or released entirely by calling this delete function.
+/// \see EnergyPlusState
+/// \see stateNew
+/// \see stateReset
+ENERGYPLUSLIB_API void stateDelete(EnergyPlusState state);
+
+#ifdef __cplusplus
+}
+#endif
 
 #endif
