+++ tst/EnergyPlus/api/TestRuntime.py
@@ -20,7 +20,7 @@ def error_handler(message: bytes) -> None:
 
 
 api = EnergyPlusAPI()
-state = api.runtime.new_state()
+state = api.state.new_state()
 api.runtime.callback_begin_new_environment(state, environment_handler)
 api.runtime.callback_progress(state, progress_handler)
 api.functional.callback_error(state, error_handler)
